+++ internal/internal_worker.go
@@ -21,6 +21,7 @@ import (
 	"github.com/google/uuid"
 	"github.com/nexus-rpc/sdk-go/nexus"
 	commonpb "go.temporal.io/api/common/v1"
+	deploymentpb "go.temporal.io/api/deployment/v1"
 	enumspb "go.temporal.io/api/enums/v1"
 	historypb "go.temporal.io/api/history/v1"
 	"go.temporal.io/api/temporalproto"
@@ -211,6 +212,18 @@ type (
 		// LastEventID, if set, will only load history up to this ID (inclusive).
 		LastEventID int64
 	}
+
+	// Represents the version of a specific worker deployment.
+	//
+	// NOTE: Experimental
+	//
+	// Exposed as: [go.temporal.io/sdk/worker.WorkerDeploymentVersion]
+	WorkerDeploymentVersion struct {
+		// The name of the deployment this worker version belongs to
+		DeploymentName string
+		// The build id specific to this worker
+		BuildId string
+	}
 )
 
 var debugMode = os.Getenv("TEMPORAL_DEBUG") != ""
@@ -259,6 +272,11 @@ func (params *workerExecutionParameters) getBuildID() string {
 		return params.WorkerBuildID
 	}
 	return getBinaryChecksum()
+}
+
+// Returns true if this worker is part of our system namespace or per-namespace system task queue
+func (params *workerExecutionParameters) isInternalWorker() bool {
+	return params.Namespace == "temporal-system" || params.TaskQueue == "temporal-sys-per-ns-tq"
 }
 
 // verifyNamespaceExist does a DescribeNamespace operation on the specified namespace with backoff/retry
@@ -322,6 +340,7 @@ func newWorkflowTaskWorkerInternal(
 		slotReservationData: slotReservationData{
 			taskQueue: params.TaskQueue,
 		},
+		isInternalWorker: params.isInternalWorker(),
 	},
 	)
 
@@ -494,6 +513,7 @@ func newActivityWorker(
 			slotReservationData: slotReservationData{
 				taskQueue: params.TaskQueue,
 			},
+			isInternalWorker: params.isInternalWorker(),
 		},
 	)
 	return &activityWorker{
@@ -1796,6 +1816,10 @@ func NewAggregatedWorker(client *WorkflowClient, taskQueue string, options Worke
 	if (options.DeploymentOptions.Version != WorkerDeploymentVersion{}) {
 		options.BuildID = options.DeploymentOptions.Version.BuildId
 	}
+	if !options.DeploymentOptions.UseVersioning &&
+		options.DeploymentOptions.DefaultVersioningBehavior != VersioningBehaviorUnspecified {
+		panic("cannot set both DeploymentOptions.DefaultVersioningBehavior if DeploymentOptions.UseBuildIDForVersioning is false")
+	}
 
 	// Need reference to result for fatal error handler
 	var aw *AggregatedWorker
@@ -2185,4 +2209,42 @@ func executeFunction(fn interface{}, args []interface{}) (interface{}, error) {
 		res = retValues[0].Interface()
 	}
 	return res, err
+}
+
+func workerDeploymentVersionFromProto(wd *deploymentpb.WorkerDeploymentVersion) WorkerDeploymentVersion {
+	return WorkerDeploymentVersion{
+		DeploymentName: wd.DeploymentName,
+		BuildId:        wd.BuildId,
+	}
+}
+
+func (wd *WorkerDeploymentVersion) toProto() *deploymentpb.WorkerDeploymentVersion {
+	return &deploymentpb.WorkerDeploymentVersion{
+		DeploymentName: wd.DeploymentName,
+		BuildId:        wd.BuildId,
+	}
+}
+
+func (wd *WorkerDeploymentVersion) toCanonicalString() string {
+	return fmt.Sprintf("%s.%s", wd.DeploymentName, wd.BuildId)
+}
+
+func workerDeploymentVersionFromString(version string) *WorkerDeploymentVersion {
+	if splitVersion := strings.SplitN(version, ".", 2); len(splitVersion) == 2 {
+		return &WorkerDeploymentVersion{
+			DeploymentName: splitVersion[0],
+			BuildId:        splitVersion[1],
+		}
+	}
+	return nil
+}
+
+func workerDeploymentVersionFromProtoOrString(wd *deploymentpb.WorkerDeploymentVersion, fallback string) *WorkerDeploymentVersion {
+	if wd == nil {
+		return workerDeploymentVersionFromString(fallback)
+	}
+	return &WorkerDeploymentVersion{
+		DeploymentName: wd.DeploymentName,
+		BuildId:        wd.BuildId,
+	}
 }
